Migrate to PostgreSQL 18 + pgvector 0.8.4
- SQLite → PostgreSQL 18 (Docker container archetype-postgres on Colima)
- pgvector 0.8.4 extension enabled
- sqlx features: sqlite → postgres
- SqlitePool → PgPool across all modules (config, state, db/queries, tests)
- Migration SQL: SERIAL PRIMARY KEY (PostgreSQL-native)
- ModelEntry.vision/tools: i64 → i32 (Postgres int4)
- 61 rows migrated from SQLite via \COPY
- launchd plist updated with postgres:// connection string
- Dockerfile.postgres: postgres:18-bookworm + pgvector apt
- Docker volume archetype-pgdata persists data
- Postgres 18 gotcha: data dir is /var/lib/postgresql NOT /var/lib/postgresql/data
- 4 integration tests passing against live PostgreSQL
- All endpoints verified: status, summary (61 rows), models (40 entries), SSE

+++ src/models/model_entry.rs
@@ -6,7 +6,7 @@ pub struct ModelEntry {
     pub name: String,
     pub provider: String,
     pub kind: String,
-    pub vision: i64,
-    pub tools: i64,
+    pub vision: i32,
+    pub tools: i32,
     pub local_path: String,
 }
